requirements_clean.txt
# MP3 Downloader - Clean Requirements
# Core dependencies only

# YouTube downloading
yt-dlp>=2024.1.0

# Telegram bot
python-telegram-bot>=20.0

# Web interface (optional)
flask>=3.0.0

# Additional utilities
pathlib2>=2.3.0; python_version < '3.4'